Write the string to use with -ss in the ITP
The secondary structure that is written in the ITP output uses an
internal representation of the types. This representation is not
compatible with the -ss option of martinize which results in a different
output when using -ss from when using -dssp when copying the secondary
structure from the ITP.

This commit makes martinize write, in the ITP file, a string that can be
used with -ss. This string is taken either from the output of DSSP, or
directly from the string that has been provided with -ss.

This addresses the issue #7.